+++ tests/src/core/getExports.js
@@ -83,11 +83,12 @@ describe('getExports', function () {
     var imports = ExportMap.parse(
       path,
       contents,
-      { parserPath: 'babel-eslint' }
+      { parserPath: 'babel-eslint', settings: {} }
     )
 
-    expect(imports).to.exist
-    expect(imports.get('default')).to.exist
+    expect(imports, 'imports').to.exist
+    expect(imports.errors).to.be.empty
+    expect(imports.get('default'), 'default export').to.exist
     expect(imports.has('Bar')).to.be.true
   })
 
@@ -187,6 +188,7 @@ describe('getExports', function () {
           sourceType: 'module',
           attachComment: true,
         },
+        settings: {},
       })
     })
 
@@ -197,6 +199,7 @@ describe('getExports', function () {
           sourceType: 'module',
           attachComment: true,
         },
+        settings: {},
       })
     })
   })
@@ -293,7 +296,7 @@ describe('getExports', function () {
   })
 
   context('issue #478: never parse non-whitelist extensions', function () {
-    const context = Object.assign({}, fakeContext,
+    const context = assign({}, fakeContext,
       { settings: { 'import/extensions': ['.js'] } })
 
     let imports
@@ -303,6 +306,37 @@ describe('getExports', function () {
 
     it('returns nothing for a TypeScript file', function () {
       expect(imports).not.to.exist
+    })
+
+  })
+
+  context('alternate parsers', function () {
+    const configs = [
+      // ['string form', { 'typescript-eslint-parser': '.ts' }],
+      ['array form', { 'typescript-eslint-parser': ['.ts', '.tsx'] }],
+    ]
+
+    configs.forEach(([description, parserConfig]) => {
+      describe(description, function () {
+        const context = assign({}, fakeContext,
+          { settings: {
+            'import/extensions': ['.js'],
+            'import/parsers': parserConfig,
+          } })
+
+        let imports
+        before('load imports', function () {
+          imports = ExportMap.get('./typescript.ts', context)
+        })
+
+        it('returns something for a TypeScript file', function () {
+          expect(imports).to.exist
+        })
+
+        it('has export (getFoo)', function () {
+          expect(imports.has('getFoo')).to.be.true
+        })
+      })
     })
 
   })